pyproject.toml
[project]
name = "auto-aim-rust"
version = "0.1.0"
description = "AutoAim for RoboMaster using Rust"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "onnx>=1.18.0",
    "onnxruntime>=1.22.1",
    "opencv-python>=4.12.0.88",
    "openvino>=2025.2.0",
    "rerun-sdk>=0.24.1",
    "scipy>=1.16.1",
]
